Added Door Unlock
Added DoorManager script to Door prefab to control door opening if player has the key.
Added isFakeDoor attribute to prefab Door so that doors that cannot be opened (e.g. entrance) will not be unlocked with a key.
Updated KeyManager script to trigger KeyCollected function in PlayerData script (to update key collection status).